requirements.txt
streamlit
pandas
numpy
altair
openpyxl
matplotlib
python-docx
supabase
passlib[bcrypt]
